requirements.txt
# === Transcrição em Tempo Real ===
faster-whisper>=1.1.0
sounddevice>=0.5.0
numpy>=1.24.0
webrtcvad>=2.0.10

# === LLM Local (Ollama) ===
ollama>=0.4.0

# === Base de Conhecimento (RAG) ===
chromadb>=0.5.0
sentence-transformers>=3.0.0
PyMuPDF>=1.23.0  # Para ler PDFs de artigos científicos

# === Interface Web ===
flask>=3.0.0
flask-socketio>=5.3.0
gevent>=24.0
gevent-websocket>=0.10.1

# === Fine-Tuning ===
# Nota: unsloth deve ser instalado separadamente
# pip install "unsloth[colab-new] @ git+https://github.com/unslothai/unsloth.git"
# pip install --no-deps trl peft accelerate bitsandbytes
datasets>=3.0.0

# === Utilitários ===
python-dotenv>=1.0.0
rich>=13.0.0
pydantic>=2.0.0
